# 4.x: Reduce N-ary StableCompositeDisposable.Create overhead via dedicated implementation