--- a/Rx.NET/Source/src/System.Reactive/Linq/Observable/CombineLatest.Generated.tt
+++ b/Rx.NET/Source/src/System.Reactive/Linq/Observable/CombineLatest.Generated.tt
@@ -96,7 +96,7 @@ for (var j = 1; j <= i; j++)
 }
 #>
 
-                SetUpstream(StableCompositeDisposable.Create(subscriptions));
+                SetUpstream(StableCompositeDisposable.CreateTrusted(subscriptions));
             }
 
             protected override TResult GetResult() => _resultSelector(<#=vs#>);