… Rx.NET/Source/System.Reactive.sln
… Rx.NET/Source/benchmarks/Benchmarks.System.Reactive/StableCompositeDisposableBenchmark.cs