This PR reduces the overhead of the N-ary `StableCompositeDisposable.Create` by splitting the implementation into 3:

- An array-based implementation to support the **public** params version
- A list-based implementation to keep supporting the **public** `IEnumerable` version. Note that Rx.NET doesn't use this overload internally but has to stay because `StableCompositeDisposable` is part of the public API.
- An array-based but *trusted internal* implementation which doesn't perform defensive copying or null-checks.

The latter relies on the fact that internal usages always present a safe, null-less array that doesn't really need copying because they are created just for the purpose of creating a fixed-size composite.

Note that most `StableCompositeDisposable` usages have only two components so the benefit is not as widespread beyond the various N-arg `CombineLatest` and `Zip` implementations.

Comparison:

```
BenchmarkDotNet=v0.10.14, OS=Windows 10.0.17134
Intel Core i7-4790 CPU 3.60GHz (Haswell), 1 CPU, 8 logical and 4 physical cores
Frequency=3513595 Hz, Resolution=284.6088 ns, Timer=TSC
  [Host]     : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3110.0
  DefaultJob : .NET Framework 4.6.2 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.3110.0
```

#### Before

Method |   N |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
-------------- |---- |----------:|----------:|----------:|-------:|----------:|
Create_Array |   3 |  76.23 ns | 0.7921 ns | 0.7409 ns | 0.0266 |     112 B |
Create_List |   3 |  41.81 ns | 0.2716 ns | 0.2408 ns | 0.0266 |     112 B |
Dispose_Array |   3 | 124.04 ns | 1.0535 ns | 0.9339 ns | 0.0265 |     112 B |
Dispose_List |   3 |  89.87 ns | 0.4768 ns | 0.4227 ns | 0.0266 |     112 B |
Create_Array |   4 |  76.39 ns | 0.2808 ns | 0.2626 ns | 0.0285 |     120 B |
Create_List |   4 |  43.73 ns | 0.3826 ns | 0.3392 ns | 0.0286 |     120 B |
Dispose_Array |   4 | 131.27 ns | 0.6209 ns | 0.5808 ns | 0.0284 |     120 B |
Dispose_List |   4 |  96.94 ns | 0.4691 ns | 0.4388 ns | 0.0285 |     120 B |
Create_Array |   5 |  76.91 ns | 0.2650 ns | 0.2349 ns | 0.0304 |     128 B |
Create_List |   5 |  43.93 ns | 0.2148 ns | 0.2009 ns | 0.0305 |     128 B |
Dispose_Array |   5 | 135.93 ns | 0.7116 ns | 0.6656 ns | 0.0303 |     128 B |
Dispose_List |   5 | 103.09 ns | 0.3232 ns | 0.2699 ns | 0.0304 |     128 B |
Create_Array |   6 |  78.88 ns | 1.0480 ns | 0.9290 ns | 0.0323 |     136 B |
Create_List |   6 |  45.82 ns | 0.5164 ns | 0.4578 ns | 0.0324 |     136 B |
Dispose_Array |   6 | 143.91 ns | 1.4902 ns | 1.3939 ns | 0.0322 |     136 B |
Dispose_List |   6 | 109.82 ns | 0.3261 ns | 0.2723 ns | 0.0323 |     136 B |
Create_Array |   7 |  78.73 ns | 0.2500 ns | 0.2338 ns | 0.0343 |     144 B |
Create_List |   7 |  45.89 ns | 0.2697 ns | 0.2390 ns | 0.0343 |     144 B |
Dispose_Array |   7 | 148.98 ns | 0.4776 ns | 0.4233 ns | 0.0343 |     144 B |
Dispose_List |   7 | 115.99 ns | 0.4064 ns | 0.3602 ns | 0.0343 |     144 B |
Create_Array |   8 |  80.97 ns | 0.3747 ns | 0.3505 ns | 0.0362 |     152 B |
Create_List |   8 |  48.29 ns | 0.1456 ns | 0.1362 ns | 0.0362 |     152 B |
Dispose_Array |   8 | 156.56 ns | 0.5811 ns | 0.4853 ns | 0.0362 |     152 B |
Dispose_List |   8 | 123.25 ns | 0.4307 ns | 0.4029 ns | 0.0362 |     152 B |
Create_Array |   9 |  81.82 ns | 0.4995 ns | 0.4672 ns | 0.0380 |     160 B |
Create_List |   9 |  48.76 ns | 0.0826 ns | 0.0773 ns | 0.0381 |     160 B |
Dispose_Array |   9 | 162.97 ns | 1.0555 ns | 0.9873 ns | 0.0379 |     160 B |
Dispose_List |   9 | 129.81 ns | 0.4291 ns | 0.3583 ns | 0.0379 |     160 B |
Create_Array |  10 |  83.68 ns | 0.5577 ns | 0.5216 ns | 0.0401 |     168 B |
Create_List |  10 |  50.12 ns | 0.2367 ns | 0.1976 ns | 0.0401 |     168 B |
Dispose_Array |  10 | 167.91 ns | 0.8301 ns | 0.7765 ns | 0.0401 |     168 B |
Dispose_List |  10 | 136.08 ns | 0.5903 ns | 0.5522 ns | 0.0401 |     168 B |
Create_Array | 100 | 202.13 ns | 1.0051 ns | 0.8910 ns | 0.2115 |     888 B |
Create_List | 100 | 168.63 ns | 0.8621 ns | 0.8065 ns | 0.2115 |     888 B |
Dispose_Array | 100 | 752.04 ns | 2.6431 ns | 2.4723 ns | 0.2108 |     888 B |
Dispose_List | 100 | 712.40 ns | 3.5192 ns | 3.2919 ns | 0.2108 |     888 B |

#### After

Method |   N |       Mean |     Error |    StdDev |  Gen 0 | Allocated |
--------------------- |---- |-----------:|----------:|----------:|-------:|----------:|
Create_Array |   3 |  38.056 ns | 0.3018 ns | 0.2823 ns | 0.0172 |      72 B |
Create_Trusted_Array |   3 |   3.635 ns | 0.0383 ns | 0.0340 ns | 0.0057 |      24 B |
Create_List |   3 |  41.883 ns | 0.1681 ns | 0.1573 ns | 0.0266 |     112 B |
Dispose_Array |   3 |  63.845 ns | 0.2786 ns | 0.2470 ns | 0.0172 |      72 B |
Dispose_List |   3 |  89.699 ns | 0.4861 ns | 0.4547 ns | 0.0266 |     112 B |
Dispose_Trused_Array |   3 |  27.021 ns | 0.1156 ns | 0.1082 ns | 0.0057 |      24 B |
Create_Array |   4 |  40.048 ns | 0.1947 ns | 0.1822 ns | 0.0190 |      80 B |
Create_Trusted_Array |   4 |   3.626 ns | 0.0284 ns | 0.0252 ns | 0.0057 |      24 B |
Create_List |   4 |  43.730 ns | 0.2188 ns | 0.2047 ns | 0.0286 |     120 B |
Dispose_Array |   4 |  67.647 ns | 0.2779 ns | 0.2600 ns | 0.0190 |      80 B |
Dispose_List |   4 |  97.430 ns | 0.4007 ns | 0.3748 ns | 0.0285 |     120 B |
Dispose_Trused_Array |   4 |  29.223 ns | 0.1636 ns | 0.1450 ns | 0.0057 |      24 B |
Create_Array |   5 |  41.814 ns | 0.2742 ns | 0.2565 ns | 0.0209 |      88 B |
Create_Trusted_Array |   5 |   3.610 ns | 0.0317 ns | 0.0281 ns | 0.0057 |      24 B |
Create_List |   5 |  43.842 ns | 0.1062 ns | 0.0994 ns | 0.0305 |     128 B |
Dispose_Array |   5 |  71.457 ns | 0.3178 ns | 0.2973 ns | 0.0209 |      88 B |
Dispose_List |   5 | 103.491 ns | 0.2819 ns | 0.2354 ns | 0.0304 |     128 B |
Dispose_Trused_Array |   5 |  31.010 ns | 0.1772 ns | 0.1571 ns | 0.0057 |      24 B |
Create_Array |   6 |  43.095 ns | 0.1779 ns | 0.1664 ns | 0.0228 |      96 B |
Create_Trusted_Array |   6 |   3.607 ns | 0.0328 ns | 0.0290 ns | 0.0057 |      24 B |
Create_List |   6 |  44.909 ns | 0.1330 ns | 0.1179 ns | 0.0324 |     136 B |
Dispose_Array |   6 |  74.614 ns | 0.2266 ns | 0.2120 ns | 0.0228 |      96 B |
Dispose_List |   6 | 110.089 ns | 0.4043 ns | 0.3782 ns | 0.0323 |     136 B |
Dispose_Trused_Array |   6 |  33.177 ns | 0.1945 ns | 0.1820 ns | 0.0057 |      24 B |
Create_Array |   7 |  44.353 ns | 0.0813 ns | 0.0721 ns | 0.0247 |     104 B |
Create_Trusted_Array |   7 |   3.638 ns | 0.0296 ns | 0.0277 ns | 0.0057 |      24 B |
Create_List |   7 |  45.632 ns | 0.1860 ns | 0.1740 ns | 0.0343 |     144 B |
Dispose_Array |   7 |  78.548 ns | 0.1800 ns | 0.1503 ns | 0.0247 |     104 B |
Dispose_List |   7 | 116.546 ns | 0.5184 ns | 0.4595 ns | 0.0343 |     144 B |
Dispose_Trused_Array |   7 |  35.530 ns | 0.1911 ns | 0.1787 ns | 0.0057 |      24 B |
Create_Array |   8 |  46.609 ns | 0.1923 ns | 0.1799 ns | 0.0267 |     112 B |
Create_Trusted_Array |   8 |   3.668 ns | 0.0267 ns | 0.0249 ns | 0.0057 |      24 B |
Create_List |   8 |  47.969 ns | 0.2628 ns | 0.2458 ns | 0.0362 |     152 B |
Dispose_Array |   8 |  82.422 ns | 0.4771 ns | 0.3984 ns | 0.0267 |     112 B |
Dispose_List |   8 | 123.191 ns | 0.5532 ns | 0.4904 ns | 0.0362 |     152 B |
Dispose_Trused_Array |   8 |  37.491 ns | 0.1760 ns | 0.1560 ns | 0.0057 |      24 B |
Create_Array |   9 |  47.421 ns | 0.1515 ns | 0.1343 ns | 0.0286 |     120 B |
Create_Trusted_Array |   9 |   3.659 ns | 0.0296 ns | 0.0277 ns | 0.0057 |      24 B |
Create_List |   9 |  48.736 ns | 0.1209 ns | 0.1131 ns | 0.0381 |     160 B |
Dispose_Array |   9 |  85.522 ns | 0.3741 ns | 0.3316 ns | 0.0285 |     120 B |
Dispose_List |   9 | 128.997 ns | 0.4867 ns | 0.4552 ns | 0.0379 |     160 B |
Dispose_Trused_Array |   9 |  39.480 ns | 0.1034 ns | 0.0967 ns | 0.0057 |      24 B |
Create_Array |  10 |  49.284 ns | 0.2822 ns | 0.2640 ns | 0.0305 |     128 B |
Create_Trusted_Array |  10 |   3.669 ns | 0.0274 ns | 0.0256 ns | 0.0057 |      24 B |
Create_List |  10 |  50.477 ns | 0.0631 ns | 0.0559 ns | 0.0401 |     168 B |
Dispose_Array |  10 |  89.544 ns | 0.3857 ns | 0.3608 ns | 0.0304 |     128 B |
Dispose_List |  10 | 136.005 ns | 0.4609 ns | 0.4085 ns | 0.0401 |     168 B |
Dispose_Trused_Array |  10 |  41.667 ns | 0.1691 ns | 0.1582 ns | 0.0057 |      24 B |
Create_Array | 100 | 161.097 ns | 0.2705 ns | 0.2398 ns | 0.2019 |     848 B |
Create_Trusted_Array | 100 |   3.729 ns | 0.0344 ns | 0.0305 ns | 0.0057 |      24 B |
Create_List | 100 | 167.805 ns | 0.8348 ns | 0.7809 ns | 0.2115 |     888 B |
Dispose_Array | 100 | 392.294 ns | 1.3666 ns | 1.2115 ns | 0.2017 |     848 B |
Dispose_List | 100 | 713.079 ns | 2.3785 ns | 2.1085 ns | 0.2108 |     888 B |
Dispose_Trused_Array | 100 | 235.087 ns | 0.6304 ns | 0.5897 ns | 0.0052 |      24 B |

#### Compare

Method |   N |      Mean |     Error |    StdDev |  Gen 0 | Allocated |
-------------- |---- |----------:|----------:|----------:|-------:|----------:|
Create_Array |   3 |  76.23 ns | 0.7921 ns | 0.7409 ns | 0.0266 |     112 B |
Create_Array_PR |   3 |  38.056 ns | 0.3018 ns | 0.2823 ns | 0.0172 |      72 B |
Create_Trusted_Array |   3 |   3.635 ns | 0.0383 ns | 0.0340 ns | 0.0057 |      24 B |
Create_List |   3 |  41.81 ns | 0.2716 ns | 0.2408 ns | 0.0266 |     112 B |
Create_List_PR |   3 |  41.883 ns | 0.1681 ns | 0.1573 ns | 0.0266 |     112 B |
Dispose_Array |   3 | 124.04 ns | 1.0535 ns | 0.9339 ns | 0.0265 |     112 B |
Dispose_Array_PR |   3 |  63.845 ns | 0.2786 ns | 0.2470 ns | 0.0172 |      72 B |
Dispose_Trused_Array |   3 |  27.021 ns | 0.1156 ns | 0.1082 ns | 0.0057 |      24 B |
Dispose_List |   3 |  89.87 ns | 0.4768 ns | 0.4227 ns | 0.0266 |     112 B |
Dispose_List_PR |   3 |  89.699 ns | 0.4861 ns | 0.4547 ns | 0.0266 |     112 B |
Create_Array |  10 |  83.68 ns | 0.5577 ns | 0.5216 ns | 0.0401 |     168 B |
Create_Array_PR |  10 |  49.284 ns | 0.2822 ns | 0.2640 ns | 0.0305 |     128 B |
Create_Trusted_Array |  10 |   3.669 ns | 0.0274 ns | 0.0256 ns | 0.0057 |      24 B |
Create_List |  10 |  50.12 ns | 0.2367 ns | 0.1976 ns | 0.0401 |     168 B |
Create_List_PR |  10 |  50.477 ns | 0.0631 ns | 0.0559 ns | 0.0401 |     168 B |
Dispose_Array |  10 | 167.91 ns | 0.8301 ns | 0.7765 ns | 0.0401 |     168 B |
Dispose_Array_PR |  10 |  89.544 ns | 0.3857 ns | 0.3608 ns | 0.0304 |     128 B |
Dispose_Trused_Array |  10 |  41.667 ns | 0.1691 ns | 0.1582 ns | 0.0057 |      24 B |
Dispose_List |  10 | 136.08 ns | 0.5903 ns | 0.5522 ns | 0.0401 |     168 B |
Dispose_List_PR |  10 | 136.005 ns | 0.4609 ns | 0.4085 ns | 0.0401 |     168 B |
Create_Array | 100 | 202.13 ns | 1.0051 ns | 0.8910 ns | 0.2115 |     888 B |
Create_Array_PR | 100 | 161.097 ns | 0.2705 ns | 0.2398 ns | 0.2019 |     848 B |
Create_Trusted_Array | 100 |   3.729 ns | 0.0344 ns | 0.0305 ns | 0.0057 |      24 B |
Create_List | 100 | 168.63 ns | 0.8621 ns | 0.8065 ns | 0.2115 |     888 B |
Create_List_PR | 100 | 167.805 ns | 0.8348 ns | 0.7809 ns | 0.2115 |     888 B |
Dispose_Array | 100 | 752.04 ns | 2.6431 ns | 2.4723 ns | 0.2108 |     888 B |
Dispose_Array_PR | 100 | 392.294 ns | 1.3666 ns | 1.2115 ns | 0.2017 |     848 B |
Dispose_Trused_Array | 100 | 235.087 ns | 0.6304 ns | 0.5897 ns | 0.0052 |      24 B |
Dispose_List | 100 | 712.40 ns | 3.5192 ns | 3.2919 ns | 0.2108 |     888 B |
Dispose_List_PR | 100 | 713.079 ns | 2.3785 ns | 2.1085 ns | 0.2108 |     888 B |