@danielcweber 2018-07-04
Do you think we should have a template that produces a StableCompositeDisposable for up to 16 arguments (or whatever the max in Zip and CombineLatest is) ?